NIFI-7429 Code review changes
