+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/api/FlowResource.java
@@ -2691,39 +2691,6 @@ public class FlowResource extends ApplicationResource {
         return generateOkResponse(entity).build();
     }
 
-    @GET
-    @Consumes(MediaType.WILDCARD)
-    @Produces(MediaType.APPLICATION_JSON)
-    @Path("node/status/history")
-    @ApiOperation(
-            value = "Gets configuration history for the node",
-            notes = NON_GUARANTEED_ENDPOINT,
-            response = ComponentHistoryEntity.class,
-            authorizations = {
-                    @Authorization(value = "Read - /flow")
-            }
-    )
-    @ApiResponses(
-            value = {
-                    @ApiResponse(code = 400, message = "NiFi was unable to complete the request because it was invalid. The request should not be retried without modification."),
-                    @ApiResponse(code = 401, message = "Client could not be authenticated."),
-                    @ApiResponse(code = 403, message = "Client is not authorized to make this request."),
-                    @ApiResponse(code = 404, message = "The specified resource could not be found."),
-                    @ApiResponse(code = 409, message = "The request was valid but NiFi was not in the appropriate state to process it. Retrying the same request later may be successful.")
-            }
-    )
-    public Response getNodeHistory() {
-        authorizeFlow();
-
-        // replicate if cluster manager
-        if (isReplicateRequest()) {
-            return replicate(HttpMethod.GET);
-        }
-
-        // generate the response
-        return generateOkResponse(serviceFacade.getNodeStatusHistory()).build();
-    }
-
     /**
      * Gets the actions for the specified component.
      *
