+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-ui/src/main/webapp/js/nf/nf-status-history.js
@@ -1176,7 +1176,7 @@
         showNodeChart: function () {
             $.ajax({
                 type: 'GET',
-                url: config.urls.api + '/flow/node/status/history',
+                url: config.urls.api + '/controller/status/history',
                 dataType: 'json'
             }).done(function (response) {
                 handleNodeStatusHistoryResponse(response.statusHistory);
